LayoutTests/fast/events/window-onerror-exception-in-attr-expected.txt

Test that window.onerror is called on window object when there is an exception(excluding syntax errors) in attribute handler. Bug 70991.

Main frame window.onerror: TypeError: null is not an object (evaluating 'null.m') at window-onerror-exception-in-attr.html:10:5 TypeError: null is not an object (evaluating 'null.m')
Main frame window.onerror: TypeError: null is not an object (evaluating 'null.m') at window-onerror-exception-in-attr.html:36:5 TypeError: null is not an object (evaluating 'null.m')
Main frame window.onerror: TypeError: null is not an object (evaluating 'null.m') at window-onerror-exception-in-attr.html:36:5 TypeError: null is not an object (evaluating 'null.m')
Button 1 Button 2 Button 3
